+++ Changelog.md
@@ -65,3 +65,4 @@
 - Fixed issue with .temp folder being deleted at the wrong time.
 - Fixed logging issues.
 - Removed unnecessary reload warn.
+- Added log message to clarify that the plugin is enabled after install.